model_armor_app/requirements.txt
Flask==2.2.5
requests==2.31.0
google-cloud-modelarmor==0.1.1

